[WIP] - add two room in stage page
