+++ src/routes/rooms/after/+page.svelte
@@ -1,15 +1,16 @@
 <script lang="ts">
-	import { createMapScene } from '$lib/room-after';
+	import { createRoomAfterScene } from '$lib/rooms';
 	import { onMount } from 'svelte';
 
 	let canvas: HTMLCanvasElement;
 
 	onMount(() => {
-		createMapScene(canvas);
+		createRoomAfterScene(canvas);
 	});
 </script>
 
 <body>
+	<a class="change-page" href="/rooms/before">See after the war</a>
 	<canvas class="webgl" id="three" bind:this={canvas}></canvas>
 </body>
 
@@ -29,5 +30,18 @@
 		top: 0;
 		left: 0;
 		outline: none;
+	}
+
+	a {
+		position: absolute;
+		bottom: 1%;
+		right: 2%;
+		z-index: 100;
+		color: black;
+		border-radius: 10px;
+		background: white;
+		font-size: 16px;
+		text-decoration: none;
+		padding: 1rem;
 	}
 </style>
